[PAN-3097] Use a BigInteger for Network ID (#1891)
Interally we use a `BigInteger` for ChainID, but currently use an `int` for
NetworkID.  Because the default for NetworkID is the same value as ChainID there
are some chains where this will be very problematic, and there is at least
one other long-living chain outside the 32 bit int value range.

This is a large commit because the type is baked fairly deep into some of
the other APIs and very much in the test code testing the type.

+++ consensus/ibftlegacy/src/main/java/tech/pegasys/pantheon/consensus/ibftlegacy/protocol/Istanbul64ProtocolManager.java
@@ -21,6 +21,7 @@ import tech.pegasys.pantheon.ethereum.p2p.rlpx.wire.Capability;
 import tech.pegasys.pantheon.ethereum.worldstate.WorldStateArchive;
 import tech.pegasys.pantheon.metrics.MetricsSystem;
 
+import java.math.BigInteger;
 import java.time.Clock;
 import java.util.List;
 
@@ -30,7 +31,7 @@ public class Istanbul64ProtocolManager extends EthProtocolManager {
   public Istanbul64ProtocolManager(
       final Blockchain blockchain,
       final WorldStateArchive worldStateArchive,
-      final int networkId,
+      final BigInteger networkId,
       final boolean fastSyncEnabled,
       final int syncWorkers,
       final int txWorkers,
@@ -54,7 +55,7 @@ public class Istanbul64ProtocolManager extends EthProtocolManager {
   public Istanbul64ProtocolManager(
       final Blockchain blockchain,
       final WorldStateArchive worldStateArchive,
-      final int networkId,
+      final BigInteger networkId,
       final boolean fastSyncEnabled,
       final int syncWorkers,
       final int txWorkers,
